New script, features, bug fixes, and removal of unnecessary "feature"
Removed the insertion of url properties from linked stylesheets as it is
unnecessary.

Added support for rewriting URLs in param and embed tags.

Fixed a minor bug with DOM based links.

Added standalone tool to just download the site file as HTML and save to
disk.  This is useful if you need to hand edit the file to host on your
phish site, use it in some larger phish, or if their HTML isn't getting
displayed quite right with gophish.